server/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
openai>=1.50.0
pydantic==2.5.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-multipart==0.0.6
bcrypt==3.2.2
passlib==1.7.1
python-jose==3.3.0
